Fix compiler warnings

+++ samples/test-matrices.cpp
@@ -259,8 +259,8 @@ TEST_CASE_TEMPLATE("matrix inverse is correct for general case", T, float,
       {1, 2, 3, 4}, {5, -6, 7, 8}, {9, 10, -11, 12}, {13, 14, 15, -16}};
   const linalg::mat<T, 4, 4> inv = inverse(mat);
   const linalg::mat<T, 4, 4> id = mul(mat, inv);
-  for (int j = 0; j < 4; ++j) {
-    for (int i = 0; i < 4; ++i) {
+  for (size_t j = 0; j < 4; ++j) {
+    for (size_t i = 0; i < 4; ++i) {
       if (i == j)
         REQUIRE(id[j][i] == doctest::Approx(1.0f));
       else
